requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
edge-tts==6.1.9
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0 